refactor(clash): move api and core manager into one mod
