+++ backend/tauri/src/core/tray/mod.rs
@@ -9,6 +9,8 @@ use super::storage;
 
 pub struct Tray {}
 
+mod proxies;
+
 impl Tray {
     pub fn tray_menu(_app_handle: &AppHandle) -> SystemTrayMenu {
         let zh = { Config::verge().latest().language == Some("zh".into()) };
@@ -134,11 +136,11 @@ impl Tray {
         #[cfg(target_os = "windows")]
         {
             let indication_icon = if *tun_mode {
-                include_bytes!("../../icons/win-tray-icon-blue.png").to_vec()
+                include_bytes!("../../../icons/win-tray-icon-blue.png").to_vec()
             } else if *system_proxy {
-                include_bytes!("../../icons/win-tray-icon-pink.png").to_vec()
+                include_bytes!("../../../icons/win-tray-icon-pink.png").to_vec()
             } else {
-                include_bytes!("../../icons/win-tray-icon.png").to_vec()
+                include_bytes!("../../../icons/win-tray-icon.png").to_vec()
             };
 
             let _ = tray.set_icon(tauri::Icon::Raw(indication_icon));
